Add a prototype, byte array optimized version of the 'Add' op
This class is of course way too specific for general use. But the point is
not to be general; The point is to have a hand-optimized op that is
supposed to have the best possible runtime on byte arrays.

The idea is to use this as a prototype for classes that are generated on
the fly, optimizing for the types of the current parameters.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>