Remove Copyable infrastructure
Functions actually do not need to know how to copy themselves, because
they are functional in nature! That is, they do not mutate themselves
when the compute(I, O) method is called. (They are mutated when they are
called as a standalone Op, but that's not how we use them in the
multithreaded mappers.) So it is actually safe to reuse the same
Function across all active threads.