Add initial support for classes as args
This mechanic allows for a Class to be passed as an argument to an op
for the purposes of better matching signatures. Sometimes you want to
pass null, but the null unfortunately provides no hints as to which type
of input you want to fill. So instead you can pass the Class and it will
still fill a null value, but with a better guess at the intended op.