Split op matching into a separate service
The op matching is complex enough that it rather pollutes the elegance
of the OpService, which is supposed to be easy to use. I wanted to
expose the findCandidates and findMatches methods, but adding them to
OpService introduces additional methods that consumers of ops will
generally not need.

So let's put those methods into a new OpMatcherService, which manages
the OperationMatchers, and let OpService go back to doing what it does
best: managing the Ops themselves.