DefaultOpMatchingService: handle optional params
This change provides a better matching of argument templates to ops with
optional parameters, by allowing any or all of the optional parameter
values to be left out of the argument list. This works by padding the
argument list with nulls for the optional parameters from right to left.

Thanks to Johannes Schindelin for pointing out that a complicated
implementation enumerating all combinations is not needed here. :-)