Generate the sources into target/generated-sources/from-template/
It is much better not to conflate generated sources with non-generated
sources. We employ the 'add-sources' goal of the build-helper plugin.

For Eclipse, happily, m2e already has support for imitating the workings
of the build-helper plugin. The support is on the m2e marketplace;
during the import of the Maven project, m2e will suggest to install it.
If it was imported previously, right-clicking on the error message will
suggest to install it, too.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>