Add a test-scope dependency on junit-benchmarks
This framework taps into JUnit and allows to make beautiful graphs and
performance regression tests based on records of previous runs.

To accomodate junit-benchmarks, we have to upgrade the junit dependency
to version 4.11.

Since upstream junit-benchmarks has a bug where it includes all @Before
and @After methods in the timing, we use our own forked version, as per
pom-scijava.

For a junit-benchmarks tutorial, see:

	http://labs.carrotsearch.com/junit-benchmarks-tutorial.html

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>