DefaultOpMatchingService: deal with invalid ops
If an op is invalid (e.g., it has an @Parameter with the "final"
keyword), it should never match anything.

We should also warn that the op is invalid in its op string.

It would be ideal to also explain *why* the op is invalid, but that task
can wait for another day...