Rename JoinFunctions and JoinInplaces
They are now called FunctionJoiner and InplaceFunctionJoiner,
respectively, and they no longer declare themselves as "join" ops.

The reason is that they are not really well-formed ops: you have to
manually pass in the functions to be joined, or else the op does not
behave as expected. So we should not expose this function joining code
at the op discovery level. (Ideally they would not even be Ops at all,
but rather only helper classes, but it is useful for them to implement
the Function interface, so the compromise is OK here.)

Since they are *not* valid "join" ops, let's rename them, both for
clarity (to avoid confusion with the JoinFunction and JoinInplace base
interfaces) and to convey that they are helper classes.