AbstractFunction: fix bug in run method
The resultant output value was never assigned back to the actual output!

In cases where the output is preallocated and passed in, this doesn't
matter, but in cases where null is passed for the output and the op
computes a new object, we need to do this. Otherwise the op execution
returns null.