MeanRealType: handle null output
I am not confident that this method of handling things is correct, or
best, but we should do _something_ when the output is null, because
function outputs are marked with required=false. So we need to create
one on the fly with some default when no explicit output is given.