Add .vm and .list for RealType implementations of arithmetic interfaces
The variables meanings can be described as:

imports - list of additional imports, see .vm for already imported packages.
javadoc - main javadoc for generated class.
interface - arithmetic interface which is implemented by this op.
authors - list of authors which will be added as javadoc @author
compute_expr - (optional) if this is set, it will be used as the expression to
	set the output to: ```output.setReal(${compute_expr});```
compute_body - (optional) if this is set, the contents of this variable will
	be inserted into the compute method.
omit_return - (optional) set to anything to not add ```return output;``` at
	the end of the compute method. This may be used if this statement al-
	ready exists in the compute_body.
constructor - Additional code to add to the contructor of this op.
members - List of members, which are represented as a map of properties:
		name - member name
		type - member type (default: double)
		modifiers - string of modifiers, e.g. "static final". Private
			is allways prepended.
		init - statement to set this member to in the class body
		construct - statement to set this member to in the class
			constructor
	All properties except name are optional.