HistogramCreate fixes
In 'HistogramCreate.run' a histogram class is created but the histogram
data does not seem to be generated.  It seems a call to 'countData' is
needed to generate the histogram of the input.

HistogramCreate is made an op so that the thresholding methods can find
it using the op service.