Add global threshold test class
This class is for testing global threshold methods.  The initial commit
has a test for Otsu threshold.  In the future tests can be added for
other global threshold methods.  The Otsu test is structured as follows:

1.  Create a ramp test image.  2.  Threshold using ops and count the
number of foreground pixels.  3.  Convert to IJ1 ImagePlus and threshold
using the ImageProcessor threshold functions 4.  Verify the count
produced by the op is the same as the count produced by ImageJ1.

imglib2-ij was added as a testing scope dependency as it contains the
utilities to convert to IJ1 ImagePlus.