Replace Create interface with CreateImg interface
* The current create interface, implies that the OpService can on basis
of the given parameters figure out what shall be created. Often this
is not the case. Imagine you simply pass a 'long[]'. The OpService can't
determine which type of Img it has to create. Actually its even unclear
that the output will be an Img.

* We have to introduce additional CreateXY interfaces, where XY is a
placeholder for certain types. CreateImg for example is inherited by ops
which create an Img given some output type and an ImgFactory.