Use autogenerated base interfaces
This migrates all the ops to use the autogenerated base interfaces of
net.imagej.ops.Ops, where applicable.

Note that in some cases, there is still an interface with the same name
in its respective package which extends the base interface -- e.g.,
net.imagej.ops.chunker.Chunker extends net.imagej.ops.Ops.Chunker. The
reason is that the base interfaces exist _only_ to provide the NAME and
optional ALIASES, as well as to offer a type-safe anchor for concrete
ops. If you want anything fancier, like an interface that defines
additional methods, and/or uses generic parameters, then it is advised
to define that interface manually in the appropriate package as well.

In other cases, the previous base interface became wholly redundant and
was hence removed in favor of the one in net.imagej.ops.Ops.