Fix compile errors in Conditions API
The AbstractCondition class now extends AbstractOutputFunction, and
knows how to create its own outputs.

One unfortunate thing in the design right now is that the output is
typed on Boolean, which is immutable. So we actually need to compute the
value of the boolean in the createOutput method, rather than in the
safeCompute method as normal.

We may want to change this design before 1.0.0, to use a boolean-style
object whose value is mutable (e.g., BitType). Otherwise, the behavior
of the Conditions API is rather unintuitive, especially if a
preallocated Boolean is passed to the op, since its value cannot change.