Introduce infrastructure for first version of feature framework
Using the feature framework, one can easily compile sets of features as
updateable functions. Compiling in this case means, that all dependencies
of features to other features and/or operations are automatically resolved
if possible. Additionally, the compilation process makes sure, that
features are not computed redundantly. For example the mean requires
the calculation of a sum as well as the sum itself. The sum is only
calculated once for both features.

the most important interfaces are:
* FeatureService: Takes care of compiling
* FeatureSet: Set of semantically similiar features
* Feature: Operation returning a double.