DefaultEquation: extend OutputFunction
The default equation op already knew how to generate its output if it
was null... but after the switch to the twin output/strict hierarchies,
it extended StrictFunction and hence could never receive a null output.
It should instead extend OutputFunction and implement the createOutput
method according to its existing logic.