ApplyConstantThreshold: avoid parallel map for now
As discussed in this thread:
http://imagej.net/pipermail/imagej-devel/2014-October/002281.html

The new BitType is not thread-safe when two cursors attempt to modify
different pixels residing in the same index of the backing LongArray.

Until that is fixed somehow, we should avoid using multithreaded code
to apply a threshold into a BitType image.

This change fixes an intermittent failure of ApplyManualThresholdTest.