OpMatchingService: highlight special parameter
If a parameter was somehow relevant to the matching failure (i.e.,
it was recorded in the OpCandidate as the status item), then let's
highlight that fact by pointing at it with a "==>" arrow.

And now that we are highlighting the relevant parameter in this way,
there is no reason to explicitly list the name in the status message.