Disable the (quite expensive) benchmarks by default
As pointed out by Curtis Rueden, benchmarks are not really useful in the
regular workflow of developing topic branches and testing commits via
regression testing.

So let's just disable the benchmark tests by default, and require the
system property imagej.ops.benchmark.tests to be set to "enabled" to run
them.

This cuts down the time spent on `mvn test` from 220 to 93 seconds on
this developer's machine.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>