Avoid recompiling generated code over and over
When the template and translation files have not changed, the exact same
code is generated as before. Due to a new timestamp, Maven wants to
recompile the generated files, though.

Let's be clever and detect when the generated files are newer than both
the corresponding template and translation files, and not bother
generating code in that case.

This cuts down the time spent on `mvn compile` after a full build from
39 to 20 seconds on this developer's machine.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>