OpMatchingService: remove isCandidate from API
There is no need right now to expose this method as public API,
especially since that signature is likely to change. Let's keep
it private until we really need it to be available externally.

I am especially glad to seal away this method since it was restricted
to CommandInfo rather than all ModuleInfo like the rest of the API.