Add a rudimentary image expression parser
For the moment, it does not handle anything fancy, just the four
arithmetic operations, ops, variables and constants.

For example, the following works:

	OOps.eval(op, "a+b*c", 1, 2, 3);

The parser is not yet clever enough to detect when temporary variables
are required, e.g. evaluating "a+gauss(a, 3)" does not work correctly
yet because the "gauss(a, 3)" should not be calculated in-place.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>