generate.groovy: use strict reference mode
We do not want the build to succeed if there are any errors in the
Velocity templates. Even if the build continued, it would almost
certainly just have compile errors immediately afterwards anyway.

In particular, this change makes it so that when a variable is not found
in the context, rather than leaving the variable reference untouched in
the generated file, the template engine throws an exception.

Fail early, fail often. Wait, no, that's not quite right... ;-)

For further reading, see:
* http://velocity.apache.org/engine/devel/user-guide.html#Strict_Reference_Mode