Unroll the net.imagej.ops.arithmetic.real classes
Their generation from templates was complex. In my opinion, the savings
in terms of lines of code was not enough to justify the increase in
complexity with respect to maintenance and ease of understanding.

The templates also contained some errors which, if left unchecked, would
have made it difficult to switch Velocity over to strict reference mode.