Convert help into a proper op
Execution of "help" is dependent on context -- i.e., which ops are
available at runtime. But eval has a similar problem, as do a few other
nested ops. The value of having these functions defined within the OPS
framework is undeniable; we'll just have to be careful how we define the
"deterministic" requirement for ops: deterministic across this context.