Add test infrastructure for type-safe op methods
We want to be sure that all op implementations which ship with the
ImageJ OPS library have corresponding type-safe methods in the
OpService (for global namespace ops), or the appropriate Namespace
subclass gateway (for non-global namespace ops).

This infrastructure will facilitate unit tests that ensure every op
implementation has one or more corresponding type-safe method
signatures. This is critical for the usability of the OPS library;
it makes discovery of what functions are available much easier,
especiallly when using an IDE with code completion such as Eclipse.