Remove AbstractCropRAI abstraction level
This abstraction level is not required anymore, especially if in the
future the OpReducer will be able to convert from RAIToRAI Ops to
ImgPlusToImgPlus operations.

IMPORTANT: We will break some API here! Test are adjusted.

However, another pass over the cropping will result in some refactoring
as we only implement the "lazy" or "view-based" version of the cropping.