Remove DimensionEqualCondition
This is a weird special case; the same could largely be accomplished
with a GetIndexOp (implementing Ops.Get) that extracts the nth index of
a list (any List, of course -- not just an ArrayList), followed by an
EqualCondition on that index.

If we somehow need this particular condition later, we can readd it.