LabelingNamespace: call run directly for built-ins
Right now, the convention is for each op's built-in method signatures to
directly invoke ops.run with the list of arguments provided, cast the
result to the known return type, and return it. This is highly likely to
be safe (the only case I can think of that causes problems is if someone
extends the concrete op implementation with another op which has
additional output parameters).

We could probably get away with calling the full method signature (i.e.,
all optional arguments specified) from the partial method signatures,
passing null for the missing optional args. But if we do this, we should
be consistent and do it across the board for all built-in methods.