Improve the completeness tests
This consolidates the completeness check for each namespace into a
single method call of a single test method. The advantages are:

* The code is now much more concise, and hence less error prone.
* Every op of the namespace is tested, without hardcoding.

One disadvantage is that if the test fails, there is no longer a
complete list of every missing method body. But adding the method bodies
and rerunning the test across multiple iterations eventually generates
all the needed information.