Remove unused Optimizer subsystem
It was added for use with the on-the-fly javassist logic, which
could synthesize optimized ops at runtime built over top other ops.

It was a cool idea, but added complexity. Now that we have a code
generation layer, we can achieve many of the same benefits, without
the mystery of runtime code synthesis.