Flesh out the "op" op, for reflectivity
It was previously called "lookup" which resulted in redundant methods
in the OpService. Now it is called "op" for succinctness and
consistency with the OpService. The only inconsistency is that the
interface and package name are "Lookup" and "lookup" respectively, to
avoid confusingly vague names which clash with the core Op interface.