Put all the op interfaces in one master class
The net.imagej.ops.Ops container class now has a static inner class
per namespace, each of which has an inner interface per op.

This makes referencing the interfaces consistent with how the ops
themselves are referenced from the OpService; for example:

  ops.math().add(...)

Is a built-in method for the ops with interface:

  Ops.Math.Add

The very picture of elegance!