DefaultOpMatchingService: match across namespaces
If an op named 'math.and' is requested, it should only match
'math.and', not 'and' ops in other namespaces such as 'logic.and'.

But if 'and' is requested, it is convenient for it to match both
'math.and' and 'logic.and'.

The worst case is that this could somehow match too many ops and
result in a "multiple 'and' ops of priority 123" error. But this
error is appropriate in such circumstances anyway.