Rename op base interface extends to use Op suffix
This is as described on the Naming page of the wiki:
https://github.com/imagej/imagej-ops/wiki/Naming

It avoids naming conflation. E.g., a base interface of Ops.Stats.Mean
is extended by net.imagej.ops.stats.mean.MeanOp. And in this way, a
separate net.imagej.ops.stats.mean.Mean might also be created to house
results of the computation if an encapsulating type is needed.

Closes #80.