Change return value of compute(I, O) to void
The social contract was that the O reference itself was always returned,
which was useful for chain calls. But it is also a point of confusion
and error in each op implemention.

If the convenience of an O return is really needed in the future, we can
create a utility method for it somewhere, to DRY up any relevant code.

There is one problem with this change: the design of OutputFunction
allows null to be passed as output to compute(I, O), and an output is
then generated via the createOutput(I) method. But the output reference
is now _thrown away_, since compute(I, O) now returns nothing. To work
around this temporarily, AbstractOutputFunction assigns the result to
its internal out variable. It would not be acceptable long-term (might
be threading issues, etc.), but is OK temporarily since a redesign of
these type hierarchies is imminent in subsequent commits. The new design
makes this problem moot, by changing the preconditions of compute(I, O)
to always disallow null for the output value.