Rename Computer -> Whizbang
This is a temporary rename; the ultimate goal is to bidirectionally
rename Computer <-> Function. However, git finds that very confusing,
so we'll use a temporary variable^H^H^H^H^H^H^H^Hname for the moment.

This is part of a series of renames for the function type hierarchy;
see #149.