Rename Function -> ComputerOp
We actually want a "function" to be compute(I) returning O, where as
a "computer" is compute(I, O) with preallocated O being passed in.
Hopefully these names are less misleading and more in line with the
usual mathematical intuition about what constitutes a function.

This is part of a series of renames for the function type hierarchy;
see #149.