Rename net.imagej.ops.join Function -> Computer
The join subpackage had classes with Function in the name, because they
operated on objects typed with the Function interface. That interface is
now called ComputerOp, so these classes should be renamed accordingly.

This commit also introduces a new JoinInplaces interface and
AbstractJoinInplaces base class for DefaultJoinInplaces, for consistency
with other join implementations in this package.

This is part of a series of renames for the function type hierarchy;
see #149.