Do not force ComputerOps to implement Converter
The Converter interface comes from ImgLib2. In the future, OPS will
split in half: scijava-ops will be the core, and imagej-ops will contain
the image-specific portion of the code, including a wealth of image
processing routines. But ComputerOp is part of the OPS core, and will
live in scijava-ops, which will not depend on imglib2.

Here we see the rule of thumb in action: prefer composition to
inheritance. In this case, we can have a ComputerConverter that adapts
any ComputerOp into a Converter, to achieve the same effect in practice.