InplaceOp: change compute(A) return value to void
The social contract was that the A reference itself was always returned,
which was useful for chain calls. But it is also a point of confusion
and error in each op implemention.

If the convenience of an A return is really needed in the future, we can
create a utility method for it somewhere, to DRY up any relevant code.

See also 28ee21ea150f7b542f11b4311aeeb7227fd351fe.