Adjust the preconditions of ComputerOp & HybridOp
Previously, it was OK to pass null to a HybridOp's compute(I, O) method.
This design was rather upside down.

This commit changes the contract such that ComputerOp#compute(I, O)
requires non-null arguments. If you want an op to support optional
output values, implement HybridOp along with the createOutput(I) method,
which will provide a default implementation of FunctionOp#compute(I)
that creates the output on the fly and passes it to
ComputerOp#compute(I, O) to populate it.

As an added benefit, this change eliminates the need for the
AbstractComputerOp#safeCompute method with protected visibility.

Closes #149!