Get rid of annoying error when building w/ Java 8
There is a test that tries to execute the example in the README.md.
That example uses the equation op, which uses javascript, and in
particular it uses the Rhino-specific importClass feature which is
not present in Java 8's Nashorn's implementation.

The workaround is to use the SciJava Javascript script language
implementation instead of the built-in vanilla javascript.

Ultimately, the scripting-javascript dependency is not desirable;
probably the equation op implementation should be changed to no
longer use javascript. But it is an acceptable short-term solution
to fix the equation op.