Move opMatcher.getOps() to ops.infos()
We are going to generalize the OpService to an OpContext, so that more
than one can exist in a given application context. One purpose of doing
that is to allow for more configurability regarding the pool of ops from
which candidates should be selected. So we need the collection of ops to
be defined in the OpService (to become OpContext), rather than the
OpMatcherService.