OpMatchingService: also consider the output types
At least, give an option for doing so when matching. You must pass the
desired output types in the OpRef using the new setOutputs method, and
then the DefaultOpMatchingService will discard candidates whose output
types do not match the ones specified.

The primary intended use case for this level of matching is to select
FunctionOp implementations based on not only their input types and
arguments, but also their special output type.