Add initialize(I) method for better batching
Unlike "normal" ops, the compute methods of ComputerOp and FunctionOp
are intended to be called repeatedly on the same op instance. As such,
there may be expensive work -- op matching especially -- that should
only be performed once and then cached, such that subsequent calls to
compute are as performant as possible.

In particular, "high-level" ops follow a frequent pattern of delegating
to "low-level" ops which have other ops as arguments. For example, a
high-level DoG op might require an image and two sigmas, which are then
used to match and configure two Gauss ops and a Subtract op, which are
then used to configure a "low-level" DoG op that simply calls those
Gauss and Subtract ops to do the work. This construction is elegant, but
to avoid negative performance implications, the low-level DoG op needs
to be matched and configured only once up front, then reused each time
the high-level DoG's compute method is called.