Shorten Input and Output accessors
Yes, it's not Java best practice. But this trick is _very_ nice to make
the code more readable in places where we access in and out, without the
lines becoming very long and/or wrapping. We already use this trick in a
few other key places, such as OpEnvironment.ops(). It is also
appropriate in this case to make the code easier to understand.