IterativeFFTFilterRAI: fix typing issue
The Java 7 compiler (1.7.0_45) didn't catch this.
But Java 8 (1.8.0_45) did:

[ERROR] .../imagej-ops/src/main/java/net/imagej/ops/filter/IterativeFFTFilterRAI.java:[168,50] incompatible types: inferred type does not conform to upper bound(s)
    inferred: O
    upper bound(s): net.imglib2.type.numeric.RealType<O>,I
[INFO] 1 error

The stats.sum method being called here has separate I and O types.
The return value will be of type O. Unfortunately, the compiler has
only limited ability to enforce the O part, since no input argument
is typed on O for that method. We want to try very hard to avoid this.
But in the meantime, we can explicitly pass the <I, O> to the call.
And of course, the sum variable needs to be typed on O, not I.