Do not use FIRST_PRIORITY for op priority
The problem with FIRST_PRIORITY is that nothing can possibly insert
itself before the op in question. Better to use VERY_HIGH_PRIORITY to
declare that the op is intended to go first under normal circumstances,
but still leave the door open for something else that REALLY wants to
have an even higher precedence.

So... when would you actually use FIRST_PRIORITY? Mostly just for
debugging. Or in certain rare circumstances (see CachedOpEnvironment).