Implement specialized initialization to convert String parameter
String parameter used to indicate MatrixOrientation isn't converted to
a MatrixOrientation object automatically by the parent abstract class.
To do so, the initialization code from the abstract class was
overridden in the feature set class, converting the one String
parameter to a MatrixOperation parameter so the correct individual
feature ops are found.

The abstract class also first called super.initialize(). Instead of
trying to call super.super.initialize(), which seems bad form and
appears to not be allowed, we copied the initialization code from the
parent of the abstract class as well.

I noticed that the feature sets were recently put into their own
package away from the abstract classes. Thus, when transferring
initialization code to the default implementations, some private
variables were needed (such as the "prioritizedOps" array and
"namedFeatureMap" object). However, probably a better way to do the
CommandService is to just add the parameter to the feature set for
context injection. Potentially though you find a way around String ->
MatrixOrientation object conversion and the initialization code can go
back into the abstract classes.