Handle special case of "" params Attribute
It seems the value of "" should be treated as "empty" but when split,
it produces an array of length 1 holding "" in it. This case is tested
for an handled now. Otherwise, the feature set abstract implementation
asks for ops with the wrong signature (i.e., including a null argument
that shouldn't exist).