CachedOpEnvironment: use simple name of delegate class instead of hash
we can't assume that the same op will be used to compute an input which
we actually want to retrieve from the cache itself. for example
two independent ops may initialize each an instance of some op A in
a cached OpEnvironment. The result should be cached (thats the whole point
of the CachedOpEnvironment).