Fix MapTest bug
Since Java converts byte to int when adding two bytes, if a byte pixel
has value 127, adding 1 to it results in an int with value of 128,
instead of a byte with value of -127. 

Therefore, test will not pass if one pixel of the test image has value
127, even if the Ops are correctly implemented. Now the unit tests
convert one comparing value to byte type when calling assertEquals.