Add delegating ops from hybrid to computer
This is mostly useful when the ComputerOp in question has a generic
type as output, which needs to be narrowed to a concrete type for the
purposes of the FunctionOp portion's return type. In this scenario, a
HybridOp cannot be used directly with type-safe generics.

For example, a ComputerOp whose output variable is a {@code T extends
RealType<T> cannot be a HybridOp because we do not know at runtime
which sort of RealType matches the caller's T parameter. However, a
separate FunctionOp can be created whose output is typed on e.g.
DoubleType, with the computation delegating to the wrapped ComputerOp.