Make FunctionViaComputers functions, not hybrids
Now that the OutputFactory is its own interface, we can have the
FunctionViaComputer extend it (and BinaryFunctionViaComputer extend
BinaryOutputFactory) without inheriting all the other aspects of
HybridOp. This is important because:

* There is no reason to use a HybridOp here since we already have the
  ComputerOp portion: it is the op being wrapped.

* In fact, it adds confusion having a HybridOp since then there are
  two different ComputerOps: the original wrapped ComputerOp, and the
  HybridOp which doubles as a ComputerOp. Not only is this harder for
  humans to sort through, it also confuses the completeness tests,
  potentially requiring extraneous method signatures in the namespace.

Better to simply have the ComputerOp class and the FunctionOp class,
which together meet all the same requirements that a HybridOp would.