Change the input type of the BinaryOp
The previous input type of BinaryOp is BinaryInput, which exposes the
"Binariness" of the Op when it works as a normal, or unary SpecialOp.
It enforces the user to create some kind of BinaryInput in order to use
it as a unary SpecialOp.

As we want the BinaryOp to be a specialization of SpecialOp, we should
make it compatible with its unary version, e.g. having the input1 as the
input, and input2 as a constant parameter.