Fix the InplaceOp type hierarchy
InplaceOp is no longer a UnaryComputerOp; it is now a separate entity
which works differently than ComputerOps do. This brings the total
number of primary special ops to three: computer, function and inplace.

Unlike computers and functions, though, an inplace op is always unary:
one single argument which is mutated.