Unify behavior of the CachedOpEnvironment
Whenever an op is requested via an OpRef, we want to wrap the result in
a cached op as appropriate (i.e., as long as it is hybrid or function).

With this change, the CachedOpEnvironment no longer needs to override
the function1 and hybrid1 methods, but rather only op(OpRef).