Centralize lookup of special ops
This adds an OpEnvironment#op(OpRef) method, which is now the
single op lookup entry point for all the *special* op lookups
(i.e., computer, function, hybrid and inplace).

Note that by design, the op(name, args) and op(type, args) methods do
_not_ feed through op(ref)! This allows the op(type, args) method in
particular to guarantee that the returned op will be an instance of the
requested type -- whereas with op(ref), it may not be so in cases where
the op needs to be adapted or wrapped -- see e.g. CustomOpEnvironment.

This also tidies up the outType parameter matching for computer and
hybrid op matching: previously null was being passed, and the output
type was implicitly constrained via the input argument types (since the
output for these ops is really a "both"). Now we explicitly specify the
output type in addition, "just in case" it might somehow matter.