WIP: Pass output type as the outType, too
--> Unfortunately, for reasons unknown, this change breaks tests. <--

This tidies up the outType parameter matching for computer, hybrid and
inplace op matching. Previously, null was being passed, and the output
type was constrained only implicitly, via the input argument types
(since the output for these ops is a "both"). Now we explicitly specify
the output type in addition, "just in case" it might somehow matter.